add key nav to Dropdown, remove index in directories because treeshake is not working this way now
