+++ packages/veui/src/components/Dropdown.vue
@@ -28,11 +28,15 @@
       ref="box"
       class="veui-dropdown-options"
       v-outside:button="close"
-      tabindex="-1">
+      tabindex="-1"
+      :ui="ui"
+      @keydown.esc.stop="expanded = false"
+      @keydown.down.prevent="navigate()"
+      @keydown.up.prevent="navigate(false)">
       <veui-option-group :options="options" ref="options">
         <slot></slot>
         <template v-if="$scopedSlots['group-label']" slot="label" slot-scope="group">
-          <slot name="group-label" v-bind="group">{{ group.label }}</slot>
+          <slot name="group-label" v-bind="group"></slot>
         </template>
         <template v-if="$scopedSlots.option" slot="option" slot-scope="option">
           <slot name="option" v-bind="option"></slot>
@@ -54,6 +58,7 @@ import OptionGroup from './Select/OptionGroup'
 import icons from '../mixins/icons'
 import overlay from '../mixins/overlay'
 import dropdown from '../mixins/dropdown'
+import keySelect from '../mixins/key-select'
 
 export default {
   name: 'veui-dropdown',
@@ -64,7 +69,7 @@ export default {
     'veui-overlay': Overlay,
     'veui-option-group': OptionGroup
   },
-  mixins: [icons, overlay, dropdown],
+  mixins: [icons, overlay, dropdown, keySelect],
   props: {
     ui: String,
     label: String,
