--- packages/veui/src/managers/index.js
@@ -1,5 +1,0 @@
-export { default as config } from './config'
-export { default as overlay } from './overlay'
-export { default as toast } from './toast'
-export { default as type } from './type'
-export { default as rule } from './rule'
