--- packages/veui/src/mixins/index.js
@@ -1,7 +1,0 @@
-export { default as ui } from './ui'
-export { default as input } from './input'
-export { default as dropdown } from './dropdown'
-export { default as form } from './form'
-export { default as table } from './table'
-export { default as icons } from './icons'
-export { default as overlay } from './overlay'
