--- packages/veui/src/plugins/index.js
@@ -1,4 +1,0 @@
-export { default as alert } from './alert'
-export { default as confirm } from './confirm'
-export { default as prompt } from './prompt'
-export { default as toast } from './toast'
